requirements.txt
vosk
sounddevice
websockets
pynput
pystray
Pillow
